requirements.txt
autopep8==1.5.7
confusable-homoglyphs==3.2.0
Django==2.2
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-decouple==2.1
django-js-asset==1.2.2
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
